Migrate chech_outdated_datasets and move commands from plugin to cli
